requirements.txt
langchain==1.1.2
langchain-openai==1.1.0
python-dotenv==1.2.1
openai==2.9.0
neo4j==6.0.3
flask==3.1.2